Refactor App component for improved flexibility
- Simplify App component code
- Add Providers component
- Replace ThemeProvider with Providers component for better flexibility
- Add optional theme prop to NextPageWithLayout type for dynamic theming
